+++ src/components/molecules/Confetti/confetti.scss
